table_scrap from pdf/requirement.txt
altgraph==0.16.1
camelot-py==0.7.3
chardet==3.0.4
Click==7.0
cycler==0.10.0
distro==1.4.0
Django==2.2.6
et-xmlfile==1.0.1
future==0.18.1
jdcal==1.4.1
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
openpyxl==3.0.0
pandas==0.25.1
pdf2image==1.9.0
pdfminer.six==20181108
pefile==2019.4.18
Pillow==6.2.0
pycryptodome==3.9.0
PyInstaller==3.5
pyparsing==2.4.2
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.3
pywin32-ctypes==0.2.0
six==1.12.0
sortedcontainers==2.1.0
sqlparse==0.3.0
tabula-py==1.4.1
XlsxWriter==1.2.2
